// src/interfaces/model/employee.interface.ts
export enum employeeStatus {
  SAVED = "SAVED",
  DRAFT = "DRAFT",
}
export interface IEmployeeCreate {
  profileType?: string;
  rollover?: boolean;
  employeeNumber?: string | number;
  rotationDate?: string | Date;
  segmentDate?: string | Date;
  salaryDate?: string | Date;
  TempNumber?: string;
  contractNumber?: string;
  timezone?: string;
  contractSignedDate?: Date;
  startDate?: Date;
  firstName: string;
  lastName: string;
  fonction?: string;
  dOB?: Date;
  placeOfBirth?: string;
  nSS?: string;
  gender: string;
  terminationDate?: Date;
  baseSalary?: number;
  travelAllowance?: number;
  Housing?: number;
  monthlySalary?: number;
  address?: string;
  medicalCheckDate?: Date;
  medicalCheckExpiry?: Date;
  medicalInsurance?: boolean;
  contractEndDate?: Date;
  LREDContractEndDate?: Date;
  dailyCost?: number;
  mobileNumber?: string;
  nextOfKinMobile?: string;
  catalogueNumber?: string;
  initialBalance?: number | null;
  photoVersionNumber?: number;
  email?: string;
  profilePicture?: string;
  clientId: number;
  segmentId?: number | null;
  subSegmentId?: number;
  employeeType?: string;
  rotationId?: number;
  residentId?: number;
  nextOfKin?: string;
  slug?: string | number;
  customBonus?: any;
  isAdminApproved?: boolean;
  isAbsenseValueInReliquat?: boolean;
  employeeStatus?: employeeStatus;
  hourlyRate?: number | null;
  fonctionDate?: Date | null;
  dateformat?: string | null;
  timeformat?: string | null;
  currency?: string | null;
  language?: string | null;
  utcstartDate?: string | null;
  utcmedicalCheckDate?: string | null;
  utcmedicalCheckExpiry?: string | null;
  utcdOB?: string | null;
  utccontractSignedDate?: string | null;
  utccontractEndDate?: string | null;
}

export interface EmployeeAttributes {
  id: number;
  loginUserId?: number;
  clientId?: number;
  employeeNumber?: string | number;
  TempNumber?: string;
  contractNumber?: string;
  contractSignedDate?: Date;
  startDate?: Date;
  fonction?: string;
  nSS?: string;
  terminationDate?: Date;
  baseSalary?: number;
  travelAllowance?: number;
  Housing?: number;
  monthlySalary?: number;
  address?: string;
  medicalCheckDate?: Date;
  medicalCheckExpiry?: Date;
  medicalInsurance?: boolean;
  contractEndDate?: Date;
  LREDContractEndDate?: Date;
  dailyCost?: number;
  hourlyRate?: number | null;
  nextOfKinMobile?: string;
  nextOfKin?: string;
  catalogueNumber?: string;
  initialBalance?: number;
  photoVersionNumber?: number;
  segmentId?: number | null;
  subSegmentId?: number;
  employeeType?: string;
  rotationId?: number;
  residentId?: number;
  customBonus?: string;
  xeroContactId?: string;
  isAdminApproved?: boolean;
  isAbsenseValueInReliquat?: boolean;
  slug?: string | number;
  oldEmployeeId?: string | null;
  employeeStatus?: employeeStatus;
  fonctionDate?: Date | null;
  createdAt?: Date | string;
  createdBy?: number;
  updatedAt?: Date | string;
  updatedBy?: number;
  deletedAt?: Date | string;
  dateformat?: string | null;
  timeformat?: string | null;
  currency?: string | null;
  bankId?: number | null;
  createdatutc?: string;
  updatedatutc?: string;
  deletedatutc?: string;
  utcstartDate?: string | null;
  utcmedicalCheckDate?: string | null;
  utcmedicalCheckExpiry?: string | null;
  utcdOB?: string | null;
  utccontractSignedDate?: string | null;
  utccontractEndDate?: string | null;
}